% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.resourcegroupstaggingapi_operations.R
\name{get_tag_keys}
\alias{get_tag_keys}
\title{Returns all tag keys in the specified region for the AWS account}
\usage{
get_tag_keys(PaginationToken = NULL)
}
\arguments{
\item{PaginationToken}{A string that indicates that additional data is available. Leave this value empty for your initial request. If the response includes a PaginationToken, use that string for this value to request an additional page of data.}
}
\description{
Returns all tag keys in the specified region for the AWS account.
}
\examples{

}
