% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.resourcegroupstaggingapi_operations.R
\name{get_tag_values}
\alias{get_tag_values}
\title{Returns all tag values for the specified key in the specified region for the AWS account}
\usage{
get_tag_values(Key, PaginationToken = NULL)
}
\arguments{
\item{Key}{The key for which you want to list all existing values in the specified region for the AWS account.}

\item{PaginationToken}{A string that indicates that additional data is available. Leave this value empty for your initial request. If the response includes a PaginationToken, use that string for this value to request an additional page of data.}
}
\description{
Returns all tag values for the specified key in the specified region for the AWS account.
}
\examples{

}
