% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.resourcegroupstaggingapi_package.R
\docType{package}
\name{paws.resourcegroupstaggingapi-package}
\alias{paws.resourcegroupstaggingapi}
\alias{paws.resourcegroupstaggingapi-package}
\title{paws.resourcegroupstaggingapi: AWS Resource Groups Tagging API}
\description{
This guide describes the API operations for the
    resource groups tagging.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
